requirements.txt
streamlit
openai-agents
python-dotenv
python-weather
